Update readme with short blurb on system user with policy (#44)
user policy is a bit preachy, updates or recommendations to improve the text are welcome.
Overall user policy on systems is to use 'admin' user. 'triplea' user has all the secrets; anyone with access to that account cannot be an honest moderator (that person can just overwrite any records in DB; hide any bans they did, frame someone else, etc). So to solve that we have a new account called 'admin'.

This account needs read-only access to database still (used to be called 'moderator')