Properly remove temp fix and do not automatically add root user keys to admin user (#63)
This change primarily fixes a 'rm -rf' that was in place, meant for a one time run and was instead left in...

This update also makes the user keys independent, so being added as a root user does not automatically grant admin access. This is done to simplify how the authorized key file is created, so it is in one place. Note that 'tasks/root_user.sh' also copies keys, with 'admin_user.sh' also creating keys we gain some consistency.

This will fix: https://github.com/triplea-game/infrastructure/issues/62: `/home/admin/` folder is being "cleaned"