Update bot version to a recent version
This is so that updates to path parameters will be handled by application code. Note, the new version number will install new files and update run scripts, but it will not otherwise alter the running application. If the bot application is restarted, then it will start up on the new version.